+++ _tests/compile-fail/test_const_generic_array_zero_length.rs
