+++ src/lib.rs
@@ -59,6 +59,8 @@
 //!
 //! Licensed under the MIT License
 
+extern crate alloc;
+
 #[macro_use]
 pub(crate) mod ringbuffer_trait;
 use core::usize;
@@ -68,9 +70,11 @@ pub use ringbuffer_trait::{RingBuffer, RingBufferExt, RingBufferRead, RingBuffer
 #[cfg(feature = "alloc")]
 mod with_alloc;
 #[cfg(feature = "alloc")]
-pub use with_alloc::AllocRingBuffer;
+pub use with_alloc::alloc_ringbuffer::AllocRingBuffer;
 #[cfg(feature = "alloc")]
-pub use with_alloc::RINGBUFFER_DEFAULT_CAPACITY;
+pub use with_alloc::alloc_ringbuffer::RINGBUFFER_DEFAULT_CAPACITY;
+#[cfg(feature = "alloc")]
+pub use with_alloc::vecdeque::GrowableAllocRingBuffer;
 
 mod with_const_generics;
 pub use with_const_generics::ConstGenericRingBuffer;
@@ -97,7 +101,8 @@ mod tests {
     use std::vec::Vec;
 
     use crate::{
-        AllocRingBuffer, ConstGenericRingBuffer, RingBuffer, RingBufferExt, RingBufferWrite,
+        AllocRingBuffer, ConstGenericRingBuffer, GrowableAllocRingBuffer, RingBuffer,
+        RingBufferExt, RingBufferRead, RingBufferWrite,
     };
 
     #[test]
@@ -113,6 +118,7 @@ mod tests {
         }
 
         test_neg_index(AllocRingBuffer::with_capacity(capacity));
+        test_neg_index(GrowableAllocRingBuffer::with_capacity(capacity));
         test_neg_index(ConstGenericRingBuffer::<usize, capacity>::new());
     }
 
@@ -124,6 +130,7 @@ mod tests {
         }
 
         test_default(AllocRingBuffer::with_capacity(8));
+        test_default(GrowableAllocRingBuffer::with_capacity(8));
         test_default(ConstGenericRingBuffer::<i32, 8>::new());
     }
 
@@ -135,11 +142,16 @@ mod tests {
         }
 
         test_new(AllocRingBuffer::with_capacity(8));
+        test_new(GrowableAllocRingBuffer::with_capacity(8));
         test_new(ConstGenericRingBuffer::<i32, 8>::new());
     }
 
     #[test]
     fn test_default_eq_new() {
+        assert_eq!(
+            GrowableAllocRingBuffer::<i32>::default(),
+            GrowableAllocRingBuffer::<i32>::new()
+        );
         assert_eq!(
             AllocRingBuffer::<i32>::default(),
             AllocRingBuffer::<i32>::new()
@@ -161,6 +173,7 @@ mod tests {
         }
 
         test_len(AllocRingBuffer::with_capacity(8));
+        test_len(GrowableAllocRingBuffer::with_capacity(8));
         test_len(ConstGenericRingBuffer::<i32, 8>::new());
     }
 
@@ -181,6 +194,16 @@ mod tests {
 
         test_len_wrap(AllocRingBuffer::with_capacity(2));
         test_len_wrap(ConstGenericRingBuffer::<i32, 2>::new());
+
+        // the growable ringbuffer actually should grow instead of wrap
+        let mut grb = GrowableAllocRingBuffer::with_capacity(2);
+        assert_eq!(0, grb.len());
+        grb.push(0);
+        assert_eq!(1, grb.len());
+        grb.push(1);
+        assert_eq!(2, grb.len());
+        grb.push(2);
+        assert_eq!(3, grb.len());
     }
 
     #[test]
@@ -196,6 +219,7 @@ mod tests {
         }
 
         test_clear(AllocRingBuffer::with_capacity(8));
+        test_clear(GrowableAllocRingBuffer::with_capacity(8));
         test_clear(ConstGenericRingBuffer::<i32, 8>::new());
     }
 
@@ -214,6 +238,7 @@ mod tests {
         }
 
         test_empty(AllocRingBuffer::with_capacity(8));
+        test_empty(GrowableAllocRingBuffer::with_capacity(8));
         test_empty(ConstGenericRingBuffer::<i32, 8>::new());
     }
 
@@ -240,6 +265,7 @@ mod tests {
         }
 
         test_iter(AllocRingBuffer::with_capacity(8));
+        test_iter(GrowableAllocRingBuffer::with_capacity(8));
         test_iter(ConstGenericRingBuffer::<i32, 8>::new());
     }
 
@@ -262,6 +288,7 @@ mod tests {
         let string = "abc".to_string();
 
         test_iter(&string, AllocRingBuffer::with_capacity(8));
+        test_iter(&string, GrowableAllocRingBuffer::with_capacity(8));
         test_iter(&string, ConstGenericRingBuffer::<&str, 8>::new());
     }
 
@@ -284,6 +311,7 @@ mod tests {
         }
 
         test_double_iter(AllocRingBuffer::with_capacity(8));
+        test_double_iter(GrowableAllocRingBuffer::with_capacity(8));
         test_double_iter(ConstGenericRingBuffer::<i32, 8>::new());
     }
 
@@ -302,6 +330,20 @@ mod tests {
 
         test_iter_wrap(AllocRingBuffer::with_capacity(2));
         test_iter_wrap(ConstGenericRingBuffer::<i32, 2>::new());
+
+        // the growable ringbuffer shouldn't actually stop growing
+        let mut b = GrowableAllocRingBuffer::with_capacity(2);
+
+        b.push(1);
+        b.push(2);
+        // No wrap
+        b.push(3);
+
+        let mut iter = b.iter();
+        assert_eq!(&1, iter.next().unwrap());
+        assert_eq!(&2, iter.next().unwrap());
+        assert_eq!(&3, iter.next().unwrap());
+        assert!(iter.next().is_none());
     }
 
     #[test]
@@ -319,6 +361,7 @@ mod tests {
         }
 
         test_iter_mut(AllocRingBuffer::with_capacity(8));
+        test_iter_mut(GrowableAllocRingBuffer::with_capacity(8));
         test_iter_mut(ConstGenericRingBuffer::<i32, 8>::new());
     }
 
@@ -338,6 +381,19 @@ mod tests {
 
         run_test_iter_mut_wrap(AllocRingBuffer::with_capacity(2));
         run_test_iter_mut_wrap(ConstGenericRingBuffer::<i32, 2>::new());
+
+        // The growable ringbuffer actually shouldn't wrap
+        let mut b = GrowableAllocRingBuffer::with_capacity(2);
+
+        b.push(1);
+        b.push(2);
+        b.push(3);
+
+        for i in b.iter_mut() {
+            *i += 1;
+        }
+
+        assert_eq!(vec![2, 3, 4], b.to_vec())
     }
 
     #[test]
@@ -358,6 +414,20 @@ mod tests {
 
         run_test_iter_mut_wrap(AllocRingBuffer::with_capacity(2));
         run_test_iter_mut_wrap(ConstGenericRingBuffer::<i32, 2>::new());
+
+        // the growable ringbuffer actually shouldn't wrap
+        let mut b = GrowableAllocRingBuffer::with_capacity(2);
+        b.push(1);
+        b.push(2);
+        b.push(3);
+
+        let buf = b.iter_mut().collect::<Vec<_>>();
+
+        for i in buf {
+            *i += 1;
+        }
+
+        assert_eq!(vec![2, 3, 4], b.to_vec())
     }
 
     #[test]
@@ -371,6 +441,7 @@ mod tests {
         }
 
         test_to_vec(AllocRingBuffer::with_capacity(8));
+        test_to_vec(GrowableAllocRingBuffer::with_capacity(8));
         test_to_vec(ConstGenericRingBuffer::<i32, 8>::new());
     }
 
@@ -387,6 +458,15 @@ mod tests {
 
         test_to_vec_wrap(AllocRingBuffer::with_capacity(2));
         test_to_vec_wrap(ConstGenericRingBuffer::<i32, 2>::new());
+
+        // The growable ringbuffer should actually remember all items
+        let mut b = GrowableAllocRingBuffer::with_capacity(2);
+
+        b.push(1);
+        b.push(2);
+        b.push(3);
+
+        assert_eq!(vec![1, 2, 3], b.to_vec())
     }
 
     #[test]
@@ -397,6 +477,7 @@ mod tests {
         }
 
         test_index(AllocRingBuffer::with_capacity(8));
+        test_index(GrowableAllocRingBuffer::with_capacity(8));
         test_index(ConstGenericRingBuffer::<i32, 8>::new());
     }
 
@@ -413,6 +494,7 @@ mod tests {
         }
 
         test_index_mut(AllocRingBuffer::with_capacity(8));
+        test_index_mut(GrowableAllocRingBuffer::with_capacity(8));
         test_index_mut(ConstGenericRingBuffer::<i32, 8>::new());
     }
 
@@ -426,6 +508,7 @@ mod tests {
         }
 
         test_peek_some(AllocRingBuffer::with_capacity(2));
+        test_peek_some(GrowableAllocRingBuffer::with_capacity(2));
         test_peek_some(ConstGenericRingBuffer::<i32, 2>::new());
     }
 
@@ -436,6 +519,7 @@ mod tests {
         }
 
         test_peek_none(AllocRingBuffer::with_capacity(8));
+        test_peek_none(GrowableAllocRingBuffer::with_capacity(8));
         test_peek_none(ConstGenericRingBuffer::<i32, 8>::new());
     }
 
@@ -460,6 +544,7 @@ mod tests {
         }
 
         test_get_relative(AllocRingBuffer::with_capacity(8));
+        test_get_relative(GrowableAllocRingBuffer::with_capacity(8));
         test_get_relative(ConstGenericRingBuffer::<i32, 8>::new());
     }
 
@@ -484,6 +569,16 @@ mod tests {
 
         test_wrapping_get_relative(AllocRingBuffer::with_capacity(2));
         test_wrapping_get_relative(ConstGenericRingBuffer::<i32, 2>::new());
+
+        // the growable ringbuffer actually shouldn't wrap
+        let mut b = GrowableAllocRingBuffer::with_capacity(2);
+        b.push(0);
+        b.push(1);
+        b.push(2);
+
+        assert_eq!(b.get(0).unwrap(), &0);
+        assert_eq!(b.get(1).unwrap(), &1);
+        assert_eq!(b.get(2).unwrap(), &2);
     }
 
     #[test]
@@ -493,6 +588,7 @@ mod tests {
         }
 
         test_get_relative_zero_length(AllocRingBuffer::with_capacity(8));
+        test_get_relative_zero_length(GrowableAllocRingBuffer::with_capacity(8));
         test_get_relative_zero_length(ConstGenericRingBuffer::<i32, 8>::new());
     }
 
@@ -516,6 +612,7 @@ mod tests {
         }
 
         test_get_relative_mut(AllocRingBuffer::with_capacity(8));
+        test_get_relative_mut(GrowableAllocRingBuffer::with_capacity(8));
         test_get_relative_mut(ConstGenericRingBuffer::<i32, 8>::new());
     }
 
@@ -542,6 +639,19 @@ mod tests {
 
         test_wrapping_get_relative_mut(AllocRingBuffer::with_capacity(2));
         test_wrapping_get_relative_mut(ConstGenericRingBuffer::<i32, 2>::new());
+
+        // the growable ringbuffer actually shouldn't wrap
+        let mut b = GrowableAllocRingBuffer::with_capacity(2);
+
+        b.push(0);
+        b.push(1);
+        b.push(2);
+
+        *b.get_mut(0).unwrap() = 3;
+
+        assert_eq!(b.get(0).unwrap(), &3);
+        assert_eq!(b.get(1).unwrap(), &1);
+        assert_eq!(b.get(2).unwrap(), &2);
     }
 
     #[test]
@@ -551,10 +661,12 @@ mod tests {
         }
 
         test_get_relative_mut_zero_length(AllocRingBuffer::with_capacity(8));
+        test_get_relative_mut_zero_length(GrowableAllocRingBuffer::with_capacity(8));
         test_get_relative_mut_zero_length(ConstGenericRingBuffer::<i32, 8>::new());
     }
 
     #[test]
+    #[allow(deprecated)]
     fn run_test_get_absolute() {
         fn test_get_absolute(mut b: impl RingBufferExt<i32>) {
             b.push(0);
@@ -584,6 +696,7 @@ mod tests {
         }
 
         test_from_iterator::<AllocRingBuffer<i32>>();
+        test_from_iterator::<GrowableAllocRingBuffer<i32>>();
         test_from_iterator::<ConstGenericRingBuffer<i32, 1024>>();
     }
 
@@ -596,6 +709,7 @@ mod tests {
         }
 
         test_from_iterator_wrap::<AllocRingBuffer<i32>>();
+        test_from_iterator_wrap::<GrowableAllocRingBuffer<i32>>();
         test_from_iterator_wrap::<ConstGenericRingBuffer<i32, 1024>>();
     }
 
@@ -634,6 +748,7 @@ mod tests {
         }
 
         test_contains(AllocRingBuffer::with_capacity(8));
+        test_contains(GrowableAllocRingBuffer::with_capacity(8));
         test_contains(ConstGenericRingBuffer::<i32, 8>::new());
     }
 
@@ -648,6 +763,7 @@ mod tests {
         }
 
         test_is_full(AllocRingBuffer::with_capacity(2));
+        test_is_full(GrowableAllocRingBuffer::with_capacity(2));
         test_is_full(ConstGenericRingBuffer::<i32, 2>::new());
     }
 
@@ -661,6 +777,7 @@ mod tests {
         }
 
         test_front_some(AllocRingBuffer::with_capacity(2));
+        test_front_some(GrowableAllocRingBuffer::with_capacity(2));
         test_front_some(ConstGenericRingBuffer::<i32, 2>::new());
     }
 
@@ -670,8 +787,9 @@ mod tests {
             assert_eq!(b.front(), None);
         }
 
-        test_front_none(AllocRingBuffer::with_capacity(8));
-        test_front_none(ConstGenericRingBuffer::<i32, 8>::new());
+        // test_front_none(AllocRingBuffer::with_capacity(8));
+        test_front_none(GrowableAllocRingBuffer::with_capacity(8));
+        // test_front_none(ConstGenericRingBuffer::<i32, 8>::new());
     }
 
     #[test]
@@ -684,6 +802,7 @@ mod tests {
         }
 
         test_back_some(AllocRingBuffer::with_capacity(2));
+        test_back_some(GrowableAllocRingBuffer::with_capacity(2));
         test_back_some(ConstGenericRingBuffer::<i32, 2>::new());
     }
 
@@ -694,6 +813,7 @@ mod tests {
         }
 
         test_back_none(AllocRingBuffer::with_capacity(8));
+        test_back_none(GrowableAllocRingBuffer::with_capacity(8));
         test_back_none(ConstGenericRingBuffer::<i32, 8>::new());
     }
 
@@ -707,6 +827,7 @@ mod tests {
         }
 
         test_front_some_mut(AllocRingBuffer::with_capacity(2));
+        test_front_some_mut(GrowableAllocRingBuffer::with_capacity(2));
         test_front_some_mut(ConstGenericRingBuffer::<i32, 2>::new());
     }
 
@@ -717,6 +838,7 @@ mod tests {
         }
 
         test_front_none_mut(AllocRingBuffer::with_capacity(8));
+        test_front_none_mut(GrowableAllocRingBuffer::with_capacity(8));
         test_front_none_mut(ConstGenericRingBuffer::<i32, 8>::new());
     }
 
@@ -730,7 +852,7 @@ mod tests {
         }
 
         test_back_some_mut(AllocRingBuffer::with_capacity(2));
-
+        test_back_some_mut(GrowableAllocRingBuffer::with_capacity(2));
         test_back_some_mut(ConstGenericRingBuffer::<i32, 2>::new());
     }
 
@@ -741,7 +863,7 @@ mod tests {
         }
 
         test_back_none_mut(AllocRingBuffer::with_capacity(8));
-
+        test_back_none_mut(GrowableAllocRingBuffer::with_capacity(8));
         test_back_none_mut(ConstGenericRingBuffer::<i32, 8>::new());
     }
 
@@ -762,6 +884,7 @@ mod tests {
         }
 
         run_test_dequeue(AllocRingBuffer::with_capacity(8));
+        run_test_dequeue(GrowableAllocRingBuffer::with_capacity(8));
         run_test_dequeue(ConstGenericRingBuffer::<i32, 8>::new());
     }
 
@@ -780,7 +903,7 @@ mod tests {
         }
 
         test_skip(AllocRingBuffer::with_capacity(8));
-
+        test_skip(GrowableAllocRingBuffer::with_capacity(8));
         test_skip(ConstGenericRingBuffer::<i32, 8>::new());
     }
 
@@ -798,6 +921,7 @@ mod tests {
         }
 
         test_skip2(AllocRingBuffer::with_capacity(2));
+        test_skip2(GrowableAllocRingBuffer::with_capacity(2));
         test_skip2(ConstGenericRingBuffer::<i32, 2>::new());
     }
 
@@ -820,6 +944,7 @@ mod tests {
         }
 
         test_push_dequeue_push(AllocRingBuffer::with_capacity(8));
+        test_push_dequeue_push(GrowableAllocRingBuffer::with_capacity(8));
         test_push_dequeue_push(ConstGenericRingBuffer::<i32, 8>::new());
     }
 
@@ -842,6 +967,7 @@ mod tests {
         }
 
         test_enqueue_dequeue_push(AllocRingBuffer::with_capacity(8));
+        test_enqueue_dequeue_push(GrowableAllocRingBuffer::with_capacity(8));
         test_enqueue_dequeue_push(ConstGenericRingBuffer::<i32, 8>::new());
     }
 
@@ -867,6 +993,26 @@ mod tests {
 
         test_push_dequeue_push_full(AllocRingBuffer::with_capacity(2));
         test_push_dequeue_push_full(ConstGenericRingBuffer::<i32, 2>::new());
+
+        // the growable ringbuffer should actually keep growing and dequeue all items
+        let mut b = GrowableAllocRingBuffer::with_capacity(2);
+        b.push(0);
+        b.push(1);
+        b.push(2);
+
+        assert_eq!(b.dequeue(), Some(0));
+        assert_eq!(b.dequeue(), Some(1));
+        assert_eq!(b.dequeue(), Some(2));
+        assert_eq!(b.dequeue(), None);
+
+        b.push(0);
+        b.push(1);
+        b.push(2);
+
+        assert_eq!(b.dequeue(), Some(0));
+        assert_eq!(b.dequeue(), Some(1));
+        assert_eq!(b.dequeue(), Some(2));
+        assert_eq!(b.dequeue(), None);
     }
 
     #[test]
@@ -899,6 +1045,35 @@ mod tests {
 
         test_push_dequeue_push_full_get(AllocRingBuffer::with_capacity(2));
         test_push_dequeue_push_full_get(ConstGenericRingBuffer::<i32, 2>::new());
+
+        // the growable ringbuffer should actually keep growing and dequeue all items
+        let mut b = GrowableAllocRingBuffer::with_capacity(2);
+
+        b.push(0);
+        b.push(1);
+        b.push(2);
+
+        assert_eq!(b.dequeue(), Some(0));
+        assert_eq!(b.dequeue(), Some(1));
+        assert_eq!(b.dequeue(), Some(2));
+        assert_eq!(b.dequeue(), None);
+
+        b.push(0);
+        b.push(1);
+        b.push(2);
+
+        assert_eq!(b.dequeue(), Some(0));
+        assert_eq!(b.dequeue(), Some(1));
+        assert_eq!(b.dequeue(), Some(2));
+        assert_eq!(b.dequeue(), None);
+
+        b.push(0);
+        b.push(1);
+        b.push(2);
+
+        assert_eq!(b.get(-1), Some(&2));
+        assert_eq!(b.get(-2), Some(&1));
+        assert_eq!(b.get(-3), Some(&0))
     }
 
     #[test]
@@ -928,6 +1103,7 @@ mod tests {
         }
 
         test_push_dequeue_push_full_get_rep(AllocRingBuffer::with_capacity(8));
+        test_push_dequeue_push_full_get_rep(GrowableAllocRingBuffer::with_capacity(8));
         test_push_dequeue_push_full_get_rep(ConstGenericRingBuffer::<i32, 8>::new());
     }
 
@@ -951,6 +1127,7 @@ mod tests {
         }
 
         test_clone(AllocRingBuffer::with_capacity(4));
+        test_clone(GrowableAllocRingBuffer::with_capacity(4));
         test_clone(ConstGenericRingBuffer::<i32, 4>::new());
     }
 
@@ -975,6 +1152,7 @@ mod tests {
         }
 
         test_default_fill(AllocRingBuffer::with_capacity(4));
+        test_default_fill(GrowableAllocRingBuffer::with_capacity(4));
         test_default_fill(ConstGenericRingBuffer::<i32, 4>::new());
     }
 
@@ -1010,6 +1188,7 @@ mod tests {
 
         next_back_test(ConstGenericRingBuffer::<i32, 8>::new());
         next_back_test(AllocRingBuffer::with_capacity(8));
+        next_back_test(GrowableAllocRingBuffer::with_capacity(8));
     }
 
     #[test]
@@ -1029,6 +1208,7 @@ mod tests {
 
         next_back_test_mut(ConstGenericRingBuffer::<i32, 8>::new());
         next_back_test_mut(AllocRingBuffer::with_capacity(8));
+        next_back_test_mut(GrowableAllocRingBuffer::with_capacity(8));
     }
     #[test]
     fn run_test_fill() {
@@ -1051,6 +1231,7 @@ mod tests {
         }
 
         test_fill(AllocRingBuffer::with_capacity(4));
+        test_fill(GrowableAllocRingBuffer::with_capacity(4));
         test_fill(ConstGenericRingBuffer::<i32, 4>::new());
     }
 
@@ -1110,6 +1291,11 @@ mod tests {
         #[test]
         fn run_test_drops_contents_const_generic() {
             test_dropped!({ ConstGenericRingBuffer::<_, 1>::new() });
+        }
+
+        #[test]
+        fn run_test_drops_contents_growable_alloc() {
+            test_dropped!({ GrowableAllocRingBuffer::with_capacity(1) });
         }
     }
 }
